implemented CreateMutex Refactoring structs.py, we dont' need sid.py or token.py
